Added handlers for the commands /help, /rules, /list
