+++ krddevbot/commands/__init__.py
@@ -1,0 +1,41 @@
+# commands/__init__.py
+
+from telegram import Update
+from telegram.ext import ContextTypes
+from ..antispam.constance import (
+    HELP_MESSAGE_TEMPLATE,
+    LIST_MESSAGE_TEMPLATE,
+    RULES_MESSAGE_TEMPLATE
+)
+from ..message_formatter import md
+
+async def delete_messages(context: ContextTypes.DEFAULT_TYPE) -> None:
+    """Deletes messages with a command and its response."""
+    chat_id = context.job.data['chat_id']
+    help_message_id = context.job.data['help_message_id']
+    
+    await context.bot.delete_message(chat_id=chat_id, message_id=help_message_id)
+    await context.bot.delete_message(chat_id=chat_id, message_id=context.job.data['command_message_id'])
+
+async def command_handler(update: Update, context: ContextTypes.DEFAULT_TYPE, *, message_template: str) -> None:
+    """Unified command handler."""
+    command_message = update.message
+    message = await command_message.reply_text(message_template)
+    
+    context.job_queue.run_once(delete_messages, when=30, data={
+        'chat_id': command_message.chat_id,
+        'help_message_id': message.message_id,
+        'command_message_id': command_message.message_id
+    })
+
+async def ping_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
+    await command_handler(update=update, context=context, message_template=md("I'm alive, {username}!", user=update.effective_user))
+
+async def help_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
+    await command_handler(update=update, context=context, message_template=HELP_MESSAGE_TEMPLATE)
+
+async def list_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
+    await command_handler(update=update, context=context, message_template=LIST_MESSAGE_TEMPLATE)
+
+async def rules_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
+    await command_handler(update=update, context=context, message_template=RULES_MESSAGE_TEMPLATE)
